split utilities for metadata into a file
